refactor: reapply directory and file name rules
