+++ src/pages/home/tab-1/Banner.tsx
